Update Start for Play 2.2
Location of the Unix script that starts a Play application has changed in  2.2. Prior, to 2.2 the Unix script was named start and it resided in the root level folder of the distribution. With 2.2 the start script is named as per the project’s name and it resides in the distribution’s bin folder.