Update stop: only kill the play process, not any other java processes.
I had the issue of git push failing due to an error in the stop script: it attempts to pkill java, however, play is not the only java process active, there is also activemq. The app has no permission to kill the activemq process so the stop script fails. These changes rectify this. 
Tested with a local Openshift Origin application.